requirements.txt
apex==0.9.10dev
cuml==0.6.1.post1
fvcore==0.1.5.post20221221
matplotlib==3.8.4
mmcv==2.0.0rc4
mmdet==3.3.0
mmsegmentation==1.2.2
numpy==2.1.2
omegaconf==2.3.0
opencv_python==4.9.0.80
opencv_python_headless==4.10.0.84
pandas==2.2.3
Pillow==10.4.0
Requests==2.32.3
scipy==1.14.1
skimage==0.0
submitit==1.5.2
torch==2.2.2
torchmetrics==1.4.3
torchvision==0.17.2
tqdm==4.65.2
transformers==4.39.3
xformers==0.0.28.post1
